Refactoring: remove redundant piece of code.
