+++ src/vm/CMakeLists.txt
@@ -329,7 +329,6 @@ set(VM_SOURCES_WKS
     eventpipebuffermanager.cpp
     eventpipesession.cpp
     eventpipesessionprovider.cpp
-    eventpipestream.cpp
     eventstore.cpp
     fastserializer.cpp
     fcall.cpp
@@ -453,7 +452,6 @@ set(VM_HEADERS_WKS
     eventpipeprovider.h
     eventpipesession.h
     eventpipesessionprovider.h
-    eventpipestream.h
     eventstore.hpp
     fastserializer.h
     fcall.h
