+++ src/vm/eventpipefile.cpp
@@ -3,15 +3,13 @@
 // See the LICENSE file in the project root for more information.
 
 #include "common.h"
-#include "eventpipebuffer.h"
 #include "eventpipeblock.h"
-#include "eventpipeconfiguration.h"
 #include "eventpipefile.h"
 #include "sampleprofiler.h"
 
 #ifdef FEATURE_PERFTRACING
 
-EventPipeFile::EventPipeFile(SString &outputFilePath) : FastSerializableObject(3, 0)
+EventPipeFile::EventPipeFile(StreamWriter *pStreamWriter) : FastSerializableObject(3, 0)
 {
     CONTRACTL
     {
@@ -39,7 +37,7 @@ EventPipeFile::EventPipeFile(SString &outputFilePath) : FastSerializableObject(3
     m_samplingRateInNs = SampleProfiler::GetSamplingRate();
 
     // Create the file stream and write the header.
-    m_pSerializer = new FastSerializer(new FileStreamWriter(outputFilePath));
+    m_pSerializer = new FastSerializer(pStreamWriter);
 
     m_serializationLock.Init(LOCK_TYPE_DEFAULT);
     m_pMetadataIds = new MapSHashWithRemove<EventPipeEvent*, unsigned int>();
@@ -62,21 +60,10 @@ EventPipeFile::~EventPipeFile()
     CONTRACTL_END;
 
     if (m_pBlock != NULL && m_pSerializer != NULL)
-    {
         WriteEnd();
-    }
 
-    if (m_pBlock != NULL)
-    {
-        delete(m_pBlock);
-        m_pBlock = NULL;
-    }
-
-    if(m_pSerializer != NULL)
-    {
-        delete(m_pSerializer);
-        m_pSerializer = NULL;
-    }
+    delete m_pBlock;
+    delete m_pSerializer;
 }
 
 bool EventPipeFile::WriteEvent(EventPipeEventInstance &instance)
@@ -102,8 +89,8 @@ bool EventPipeFile::WriteEvent(EventPipeEventInstance &instance)
 
         SaveMetadataId(*instance.GetEvent(), metadataId);
 
-        delete[] (pMetadataInstance->GetData());
-        delete (pMetadataInstance);
+        delete[] pMetadataInstance->GetData();
+        delete pMetadataInstance;
     }
 
     WriteToBlock(instance, metadataId);
@@ -143,9 +130,7 @@ void EventPipeFile::WriteToBlock(EventPipeEventInstance &instance, unsigned int 
     instance.SetMetadataId(metadataId);
 
     if (m_pBlock->WriteEvent(instance))
-    {
         return; // the block is not full, we added the event and continue
-    }
 
     // we can't write this event to the current block (it's full)
     // so we write what we have in the block to the serializer
@@ -208,9 +193,7 @@ void EventPipeFile::SaveMetadataId(EventPipeEvent &event, unsigned int metadataI
     // If a pre-existing metadata label exists, remove it.
     unsigned int oldId;
     if(m_pMetadataIds->Lookup(&event, &oldId))
-    {
         m_pMetadataIds->Remove(&event);
-    }
 
     // Add the metadata label.
     m_pMetadataIds->Add(&event, metadataId);
